Type of variance can be specified via script arg
Added a script argument '-v, --variance' to decide whether the local or global variance should
be used for computing the affinities. This is an optional argument. By
default the local variance is used. When passing '-v 0' the script will
use the global variance. For erverything else, the script will use the
local variance associcated with the given dataset.