Adapting variance normalization factor when using 3d motion vectors
since the motion variance for two given trajectoires a,b (random
variables a,b) are uncorrelected, the variance var(a+b) is equal to
var(a) + var(b). Since taking the min(var(a), var(b)) yields a
oversegmentation, this is a nice trick to avoid this particular issue.
In addition, since we work in meter scales, we have to re-scale the
variance epsilon value by 1 over the focal length squared.