Better calibration file parsing, bugfix in warping
+ only warp, if there is all necessary calibration data provided: focal
length and the prinipal point, of both, the depth-and color camera is
given. otherwise skip the warping step. In addition, zero depth values
are skipped in the computation, since they will be also zero valued at
their warped location.

+ better calibration parsing: introduced a easy kind of format which makes
use of labels. This follows the idea of yaml markup-ed files. labels are
parsed according to their value within the CalibrationManager. In
addition, the manager now assign approprate default calibration data
(which does not affect depth and tracked locations). The tracker knows
about whether rgb or/and depth intrinsic and the extrinsic parameters
were loaded.