Bugfix: Nullpointer when checking for warping required has been fixed
this error occurred when running similarity task 1 or 2 for datasets
that do not exhibit depth fields. Then, no calibration data is given
either. Therefore, the CalibrationMangager singleton was null. But we
were still invoking a method which required that singleton to exist.

An appropriate handling has been added to that particular check-method.