Bugfix: Entries of similarity matrix was not correctly sorted
A Java HashMap does not naturally sort its entry by their keys. However,
this is required by the pipelines assumpitons (label mappings).
Therefore, we have to make use of another datastructure, the so called
TreeMap with sorts the its values (asc) by its HashKey (which is the
trajectory label in our case).