Fully tested and documented SimilarityTaskType class plus refactorings
removed unused code which can be considered as legacy (the alternative
task logic). In an earlier stage of the code base, it was necessary to
provide a flag as user argument whether or not depth cues should be
used. This has been removed and every similarity task now knows whether
or not it wants to make use of depth cues. Thus, the behaviour of
falling and not falling back to another task it correct.